third_party/WebKit/LayoutTests/fast/events/drag-and-drop-autoscroll-inner-frame-expected.txt

CONSOLE WARNING: Resource requests whose URLs contain raw newline characters are deprecated, and may be blocked in M60, around August 2017. Please remove newlines from places like element attribute values in order to continue loading those resources. See https://www.chromestatus.com/features/5735596811091968 for more details.
Check autoscroll within an inner frame by drag-and-drop

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Autoscroll should have scrolled the iframe downwards, and did.
PASS iframe.contentDocument.scrollingElement.scrollTop < middleTermScrollOffset is true
PASS successfullyParsed is true

TEST COMPLETE
For manual testing, drag and drop "Drop Me" downwards and then upwards.
